requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-base-url==2.0.0
pytest-dotenv==0.5.2
jsonschema==4.20.0
Faker==21.0.0
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
allure-pytest==2.13.2
pytest-asyncio==0.21.1